requirements.txt
# Gandiva Rail Safety Monitor - Requirements
# Python 3.10 or higher required

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Serial Communication
pyserial>=3.5
pymodbus>=3.5.0

# Utilities
python-dateutil>=2.8.2
